Extract hooks from context
Since these hooks should only be used by the context, they were moved to
a subdirectory inside the store
